workspace check of files
